Idl Parsing (#1)
Complete IDL Parsing, I followd the
[specification](https://avro.apache.org/docs/1.12.0/idl-language/) and
added tests for each case. It's very possible that at this point bugs
exist, but at least the majority of cases is working.

I'm also not sure in some of the namings and workarounds I used, but for
now it seems to be good enough.